third_party/blink/web_tests/wpt_internal/attribution-reporting/register-source-disabled-by-permission-policy.https.html

<!doctype html>
<title>window.attributionReporting.registerSource: disabled by permissions policy</title>
<link rel="help" href="https://github.com/WICG/conversion-measurement-api/blob/main/event_attribution_reporting.md#registering-attribution-sources-with-javascript">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(t => {
  return promise_rejects_dom(t, 'NotAllowedError',
    window.attributionReporting.registerSource('https://a.example'));
}, 'disabled by permissions policy');

</script>
